services/market-data/backend/requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
click==8.1.7
typing-extensions==4.9.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
PyJWT==2.8.0

# Redis Caching
redis==5.0.1
cachetools==5.3.2

# Data Processing
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
scipy==1.11.4
yfinance==0.2.35

# HTTP Client
httpx==0.25.2
requests==2.31.0
certifi==2025.7.14
aiohttp==3.9.1

# Payment Processing
stripe==7.9.0

# File Processing
python-magic==0.4.27
openpyxl==3.1.2
reportlab==4.0.8

# Production Server
gunicorn==21.2.0

# Monitoring & Logging
python-json-logger==2.0.7
prometheus-client==0.19.0
ddtrace>=1.18.0
datadog>=0.47.0

# Performance
orjson==3.9.10

# Environment Management
python-dotenv==1.0.0

# AWS
boto3==1.34.14

# Template Engine
jinja2==3.1.3
markdown==3.5.1

# Async Support
aiofiles==23.2.1

# Task Queue (if needed later)
# celery==5.3.4
schedule==1.2.0

# Email (if needed)
# sendgrid==6.11.0

# Production optimizations
uvloop==0.19.0
httptools==0.6.1

# Health checks
psutil==5.9.6